+++ src/Twilio/Rest/Api/V2010/Account/Call/StreamOptions.cs
@@ -31,22 +31,22 @@ namespace Twilio.Rest.Api.V2010.Account.Call
         ///<summary> The SID of the [Call](https://www.twilio.com/docs/voice/api/call-resource) the Stream resource is associated with. </summary> 
         public string PathCallSid { get; }
 
-        ///<summary> Relative or absolute url where WebSocket connection will be established. </summary> 
+        ///<summary> Relative or absolute URL where WebSocket connection will be established. </summary> 
         public Uri Url { get; }
 
         ///<summary> The SID of the [Account](https://www.twilio.com/docs/iam/api/account) that created this Stream resource. </summary> 
         public string PathAccountSid { get; set; }
 
-        ///<summary> The user-specified name of this Stream, if one was given when the Stream was created. This may be used to stop the Stream. </summary> 
+        ///<summary> The user-specified name of this Stream, if one was given when the Stream was created. This can be used to stop the Stream. </summary> 
         public string Name { get; set; }
 
         
         public StreamResource.TrackEnum Track { get; set; }
 
-        ///<summary> Absolute URL of the status callback. </summary> 
+        ///<summary> Absolute URL to which Twilio sends status callback HTTP requests. </summary> 
         public Uri StatusCallback { get; set; }
 
-        ///<summary> The http method for the status_callback (one of GET, POST). </summary> 
+        ///<summary> The HTTP method Twilio uses when sending `status_callback` requests. Possible values are `GET` and `POST`. Default is `POST`. </summary> 
         public Twilio.Http.HttpMethod StatusCallbackMethod { get; set; }
 
         ///<summary> Parameter name </summary> 
@@ -646,7 +646,7 @@ namespace Twilio.Rest.Api.V2010.Account.Call
 
         /// <summary> Construct a new CreateStreamOptions </summary>
         /// <param name="pathCallSid"> The SID of the [Call](https://www.twilio.com/docs/voice/api/call-resource) the Stream resource is associated with. </param>
-        /// <param name="url"> Relative or absolute url where WebSocket connection will be established. </param>
+        /// <param name="url"> Relative or absolute URL where WebSocket connection will be established. </param>
         public CreateStreamOptions(string pathCallSid, Uri url)
         {
             PathCallSid = pathCallSid;
@@ -1484,7 +1484,7 @@ namespace Twilio.Rest.Api.V2010.Account.Call
         ///<summary> The SID of the [Call](https://www.twilio.com/docs/voice/api/call-resource) the Stream resource is associated with. </summary> 
         public string PathCallSid { get; }
 
-        ///<summary> The SID of the Stream resource, or the `name` used when creating the resource </summary> 
+        ///<summary> The SID or the `name` of the Stream resource to be stopped </summary> 
         public string PathSid { get; }
 
         
@@ -1497,7 +1497,7 @@ namespace Twilio.Rest.Api.V2010.Account.Call
 
         /// <summary> Construct a new UpdateStreamOptions </summary>
         /// <param name="pathCallSid"> The SID of the [Call](https://www.twilio.com/docs/voice/api/call-resource) the Stream resource is associated with. </param>
-        /// <param name="pathSid"> The SID of the Stream resource, or the `name` used when creating the resource </param>
+        /// <param name="pathSid"> The SID or the `name` of the Stream resource to be stopped </param>
         /// <param name="status">  </param>
         public UpdateStreamOptions(string pathCallSid, string pathSid, StreamResource.UpdateStatusEnum status)
         {