+++ src/Twilio/Rest/FlexApi/V2/WebChannelsResource.cs
@@ -45,7 +45,7 @@ namespace Twilio.Rest.FlexApi.V2
                 Rest.Domain.FlexApi,
                 path,
                 postParams: options.GetParams(),
-                headerParams: null
+                headerParams: options.GetHeaderParams()
             );
         }
 
@@ -79,6 +79,7 @@ namespace Twilio.Rest.FlexApi.V2
         /// <param name="chatFriendlyName"> The Conversation's friendly name. See the [Conversation resource](https://www.twilio.com/docs/conversations/api/conversation-resource) for an example. </param>
         /// <param name="customerFriendlyName"> The Conversation participant's friendly name. See the [Conversation Participant Resource](https://www.twilio.com/docs/conversations/api/conversation-participant-resource) for an example. </param>
         /// <param name="preEngagementData"> The pre-engagement data. </param>
+        /// <param name="uiVersion"> The Ui-Version HTTP request header </param>
         /// <param name="client"> Client to make requests to Twilio </param>
         /// <returns> A single instance of WebChannels </returns>
         public static WebChannelsResource Create(
@@ -86,9 +87,10 @@ namespace Twilio.Rest.FlexApi.V2
                                           string chatFriendlyName = null,
                                           string customerFriendlyName = null,
                                           string preEngagementData = null,
+                                          string uiVersion = null,
                                           ITwilioRestClient client = null)
         {
-            var options = new CreateWebChannelsOptions(addressSid){  ChatFriendlyName = chatFriendlyName, CustomerFriendlyName = customerFriendlyName, PreEngagementData = preEngagementData };
+            var options = new CreateWebChannelsOptions(addressSid){  ChatFriendlyName = chatFriendlyName, CustomerFriendlyName = customerFriendlyName, PreEngagementData = preEngagementData, UiVersion = uiVersion };
             return Create(options, client);
         }
 
@@ -98,6 +100,7 @@ namespace Twilio.Rest.FlexApi.V2
         /// <param name="chatFriendlyName"> The Conversation's friendly name. See the [Conversation resource](https://www.twilio.com/docs/conversations/api/conversation-resource) for an example. </param>
         /// <param name="customerFriendlyName"> The Conversation participant's friendly name. See the [Conversation Participant Resource](https://www.twilio.com/docs/conversations/api/conversation-participant-resource) for an example. </param>
         /// <param name="preEngagementData"> The pre-engagement data. </param>
+        /// <param name="uiVersion"> The Ui-Version HTTP request header </param>
         /// <param name="client"> Client to make requests to Twilio </param>
         /// <returns> Task that resolves to A single instance of WebChannels </returns>
         public static async System.Threading.Tasks.Task<WebChannelsResource> CreateAsync(
@@ -105,9 +108,10 @@ namespace Twilio.Rest.FlexApi.V2
                                                                                   string chatFriendlyName = null,
                                                                                   string customerFriendlyName = null,
                                                                                   string preEngagementData = null,
+                                                                                  string uiVersion = null,
                                                                                   ITwilioRestClient client = null)
         {
-        var options = new CreateWebChannelsOptions(addressSid){  ChatFriendlyName = chatFriendlyName, CustomerFriendlyName = customerFriendlyName, PreEngagementData = preEngagementData };
+        var options = new CreateWebChannelsOptions(addressSid){  ChatFriendlyName = chatFriendlyName, CustomerFriendlyName = customerFriendlyName, PreEngagementData = preEngagementData, UiVersion = uiVersion };
             return await CreateAsync(options, client);
         }
         #endif