+++ CHANGES.md
@@ -1,6 +1,18 @@
 twilio-csharp Changelog
 =======================
 
+[2024-04-18] Version 7.0.5
+--------------------------
+**Flex**
+- Add header `ui_version` to `web_channels` API
+
+**Messaging**
+- Redeploy after failed pipeline
+
+**Numbers**
+- Add Delete Port In request phone number api and Add Delete Port In request api
+
+
 [2024-04-04] Version 7.0.4
 --------------------------
 **Api**